Manually update instances
The nightly CI has been broken due to a problem with SimplyTranslate,
which hasn't been fixed for some time. This will be fixed (on Farside's
end) shortly, but the CI build should also be updated to skip instance
update routines if there's a connection error or an empty list is
returned.